requirements.txt
asgiref==3.7.2
Babel==2.14.0
Django==5.0.1
django-cors-headers==4.3.1
django-money==3.4.1
djangorestframework==3.14.0
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
py-moneyed==3.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
